Data Generators

Data Generators

The package offers (currently) two data generators, one for Multinomial priors, and the other for Gaussians.

DPMMSubClusters.generate_gaussian_data — Function.

generate_gaussian_data(N::Int64, D::Int64, K::Int64,MixtureVar::Number)

Generate N observations, generated from K D dimensions Gaussians, with the Gaussian means sampled from a Normal distribution with mean 0 and MixtureVar variance.

Returns (Samples, Labels, Clusters_means, Clusters_cov)

Example

julia> x,y,clusters = generate_gaussian_data(10000,2,6,100.0)
[3644, 2880, 119, 154, 33, 3170]
...

source

DPMMSubClusters.generate_mnmm_data — Function.

 generate_mnmm_data(N::Int64, D::Int64, K::Int64, trials::Int64)

Generate N observations, generated from K D features Multinomial vectors, with trials draws from each vector.

Returns (Samples, Labels, Vectors)

Example

julia> generate_mnmm_data(10000, 10, 5, 100)
...

source